+++ tests/unit/database/query/compilers/BaseCompilerTest.php
@@ -293,6 +293,21 @@ class BaseCompilerTest extends TestCase
 	/**
 	 *
 	 */
+	public function testSelectWithSharedLockMethod()
+	{
+		$query = $this->getBuilder();
+
+		$query->sharedLock();
+
+		$query = $query->getCompiler()->select();
+
+		$this->assertEquals('SELECT * FROM "foobar"', $query['sql']);
+		$this->assertEquals([], $query['params']);
+	}
+
+	/**
+	 *
+	 */
 	public function testSelectWithCustomLock()
 	{
 		$query = $this->getBuilder();