+++ tests/unit/database/query/compilers/DB2CompilerTest.php
@@ -119,6 +119,21 @@ class DB2CompilerTest extends TestCase
 	/**
 	 *
 	 */
+	public function testSelectWithSharedLockMethod()
+	{
+		$query = $this->getBuilder();
+
+		$query->sharedLock();
+
+		$query = $query->getCompiler()->select();
+
+		$this->assertEquals('SELECT * FROM "foobar" FOR READ ONLY WITH RS', $query['sql']);
+		$this->assertEquals([], $query['params']);
+	}
+
+	/**
+	 *
+	 */
 	public function testSelectWithCustomLock()
 	{
 		$query = $this->getBuilder();