+++ tests/unit/database/query/compilers/FirebirdCompilerTest.php
@@ -105,6 +105,21 @@ class FirebirdCompilerTest extends TestCase
 	/**
 	 *
 	 */
+	public function testSelectWithSharedLockMethod()
+	{
+		$query = $this->getBuilder();
+
+		$query->sharedLock();
+
+		$query = $query->getCompiler()->select();
+
+		$this->assertEquals('SELECT * FROM "foobar" WITH LOCK', $query['sql']);
+		$this->assertEquals([], $query['params']);
+	}
+
+	/**
+	 *
+	 */
 	public function testSelectWithCustomLock()
 	{
 		$query = $this->getBuilder();