+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Update using ```composer update```.
 * Added `BelongsToPolymorphic` relation.
 * Added `ORM::$isPersisted` property.
 * Added `ORM::isPersisted()` method.
+* Added `Query::sharedLock()` convenience method.
 * It is now possible to use aggregate methods (`count`, `min`, `max`, etc...) in a subquery context.
 * Migrations are now executed in a transaction if possible (Postgres, SQLite).
 * Added multi database support to migrations.